refactor:フックの呼び出しを移動
